feat(admin-api): service-token + blog-audio endpoints; move admin UI out of the web app
The service-token and blog-audio admin belong in the octopus-configuration
console (the vendor admin surface), not as session-gated pages in the public
web app — those 404'd for the vendor and split admin across two apps.

- Add ADMIN_API_SECRET-gated endpoints (same pattern as /api/admin/seed-docs
  & telemetry) that the config app calls via lib/octopus-api.ts:
  - GET/POST /api/admin/service-tokens (list + mint, reveal-once) and
    DELETE /api/admin/service-tokens/[id] (revoke). Reuses lib/scopes +
    lib/api-auth so the hash/scopes can't drift from /api/blog auth.
  - GET/POST /api/admin/blog-audio (state; save-voice | generate | clear).
    Reuses lib/elevenlabs + lib/r2; generate cache-busts the audioUrl.
- Remove the orphaned /admin/tokens and /admin/blog-audio pages from apps/web.
  Kept lib/elevenlabs, lib/scopes, lib/api-auth, the /api/blog scope
  enforcement, the mint-service-token script, and SystemConfig.blogAudioVoiceId
  (all still used at runtime).

The matching UI is added in octopus-configuration in a follow-up.
